trigger/requirements.txt
functions-framework==3.*
google-cloud-run==0.10.*
cloudevents==1.11.*
